Janky/Dirty fix for issue #115
Client-side only change, as it only changes the GUICorpse.java file, which is client-side only.

Allows the player to continue to click through pages, up to page 200 no matter what data the server sends about how many slots are in corpse.getSizeInventory().

This is done because once a corpse is unloaded/reloaded, the server will always tell the client there is only 54 slots (1 page). I wasn't able to find where exactly this logic is to fix the root issue, and honestly I don't care to as this solution works fine for me.

If you don't implement this change, I would like to request permission to upload this recompiled version of the 1.12.2 corpse mod to Curse myself so others can access a fixed version without having to re-compile themselves like I did.